==> 02-NoteBooks/01-Baseline.ipynb <==
# Libraries

import pandas as pd
import numpy as np
from sklearn.metrics import mean_squared_log_error

pd.set_option('display.max_columns', 200)

# Read Data

data = pd.read_csv('../01-Data/2004-2019.tsv', sep='\t', parse_dates=['DATA INICIAL', 'DATA FINAL'])

data.head()

# Translate column names to English

data.columns

data.columns = [
                    "Unnamed:_0", 
                    "Analysis_Date", #DATA INICIAL
                    "Last_Day_of_Analyses_of_Week", #DATA FINAL
                    "Macroregion", #REGIÃO
                    "State", #ESTADO
                    "Product", #PRODUTO
                    "No_of_Gas_Stations_Analyzed", #NÚMERO DE POSTOS PESQUISADOS
                    "Measurement_Unit", #UNIDADE DE MEDIDA
                    "Mean_Price", #PREÇO MÉDIO REVENDA
                    "Std_Dev", #MARGEM MÉDIA REVENDA
                    "Min_Price", #DESVIO PADRÃO REVENDA
                    "Max_Price", #PREÇO MÁXIMO REVENDA
                    "Mean_Price_Margin", #MARGEM MÉDIA REVENDA
                    "Coefficient_of_variation", #COEF DE VARIAÇÃO REVENDA
                    "Mean_Dist_Price", #PREÇO MÉDIO DISTRIBUIÇÃO
                    "Distribution_Std_Dev", #DESVIO PADRÃO DISTRIBUIÇÃO
                    "Distribution_Min_Price", #PREÇO MÍNIMO DISTRIBUIÇÃO
                    "Distribution_Max_Price", #PREÇO MÁXIMO DISTRIBUIÇÃO
                    "Distribution_Coefficient_of_Variation", #COEF DE VARIAÇÃO DISTRIBUIÇÃO
                    "Month", #MÊS
                    "Year" #ANO
]

data.columns

# Date Value Check

(data['Last_Day_of_Analyses_of_Week'] - data['Analysis_Date']).value_counts()

#### All date ranges are valid:
- end date is greater than start date

# Determination of the baseline

## dataset segmentation for gasoline

data = data[data['Product'] == 'GASOLINA COMUM']

## Train and Validation Split (Simple Holdout)

data_train = data[data['Last_Day_of_Analyses_of_Week'] < '2011-01-01']
data_valid = data[data['Last_Day_of_Analyses_of_Week'] >= '2011-01-01']

data_train.shape, data_valid.shape

y_train = data_train['Mean_Price']
y_valid = data_valid['Mean_Price']

## Baseline

baseline_train = data_train.groupby(['State'])['Mean_Price'].shift(1)
baseline_valid = data_valid.groupby(['State'])['Mean_Price'].shift(1)

metric = np.sqrt(mean_squared_log_error(y_valid[baseline_valid.notnull()],
                                      baseline_valid[baseline_valid.notnull()]))

print(metric * 100)

##### the logarithmic average square metric of the error was adopted because it is close to the percentage error

# Convert Data to .csv

data.to_csv('../01-Data/DataGas.csv',index=False)
