==> 02-NoteBooks/03-FirstModels.ipynb <==
# Libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import warnings

warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', 200)

from sklearn.metrics import mean_squared_log_error

from sklearn.ensemble import RandomForestRegressor
from lightgbm import LGBMRegressor
from sklearn.linear_model import LinearRegression
from sklearn.linear_model import Lasso
from sklearn.linear_model import Ridge
from sklearn.linear_model import BayesianRidge

# Read Data

data = pd.read_csv('../01-Data/DataGas.csv', parse_dates=['Analysis_Date', 'Last_Day_of_Analyses_of_Week'])

data.head()

data.columns

# Train and Validation Split (Simple Holdout)

data_train = data[data['Last_Day_of_Analyses_of_Week'] < '2011-01-01']
data_valid = data[data['Last_Day_of_Analyses_of_Week'] >= '2011-01-01']

data_train.shape, data_valid.shape

# New DataFrame for Train and Validation (Index: original Data)

df_train = pd.DataFrame(index=data_train.index)
df_valid  = pd.DataFrame(index=data_valid.index)

# Target

## First Difference of Average Resale Price

df_train['diff_Mean_Price'] = data_train.groupby(['State'])['Mean_Price'].apply(lambda row: row.diff().shift(-1))
df_valid['diff_Mean_Price'] = data_valid.groupby(['State'])['Mean_Price'].apply(lambda row: row.diff().shift(-1))

# Features

## Current Mean Price

df_train['Current_Mean_Price'] = data_train['Mean_Price']
df_valid['Current_Mean_Price'] = data_valid['Mean_Price']

## Seasonality

df_train['month'] = data_train['Last_Day_of_Analyses_of_Week'].dt.month
df_train['day'] = data_train['Last_Day_of_Analyses_of_Week'].dt.day
df_train['weekday'] = data_train['Last_Day_of_Analyses_of_Week'].dt.weekday
df_train['dayofyear'] = data_train['Last_Day_of_Analyses_of_Week'].dt.dayofyear

df_valid['month'] = data_valid['Last_Day_of_Analyses_of_Week'].dt.month
df_valid['day'] = data_valid['Last_Day_of_Analyses_of_Week'].dt.day
df_valid['weekday'] = data_valid['Last_Day_of_Analyses_of_Week'].dt.weekday
df_valid['dayofyear'] = data_valid['Last_Day_of_Analyses_of_Week'].dt.dayofyear

## Movel Average

roll = data_train.groupby(['State'])['Mean_Price'].rolling(4).mean()
roll.reset_index(level=[0,1], drop=True).head()

df_train['Movel_Average_Mean_Price_4_weeks'] = data_train.groupby(['State'])['Mean_Price'].rolling(4).mean().reset_index(level=0, drop=True)
df_valid['Movel_Average_Mean_Price_4_weeks'] = data_valid.groupby(['State'])['Mean_Price'].rolling(4).mean().reset_index(level=0, drop=True)

df_train.isnull().sum()

df_train = df_train.dropna()
df_valid = df_valid.dropna()

## X, y Train and Validation Split

Xtr, ytr = df_train.drop(['diff_Mean_Price'], axis=1), df_train['diff_Mean_Price']
Xval, yval = df_valid.drop(['diff_Mean_Price'], axis=1), df_valid['diff_Mean_Price']

# Model

## RandomForestRegressor

mdl = RandomForestRegressor(n_jobs=-1, random_state=0, n_estimators=500)
mdl.fit(Xtr, ytr)
p = mdl.predict(Xval)

#### Metric

p_final = Xval['Current_Mean_Price'] + p
yval_final = Xval['Current_Mean_Price'] + yval

np.sqrt(mean_squared_log_error(yval_final, p_final)) * 100

## LGBMRegressor

mdl = LGBMRegressor(num_leaves=2, min_data_in_leaf=250, n_jobs=-1, random_state=0, n_estimators=500)
mdl.fit(Xtr, ytr)
p = mdl.predict(Xval)

#### Metric

p_final = Xval['Current_Mean_Price'] + p
yval_final = Xval['Current_Mean_Price'] + yval

np.sqrt(mean_squared_log_error(yval_final, p_final)) * 100

## LinearRegression

mdl = LinearRegression()
mdl.fit(Xtr, ytr)
p = mdl.predict(Xval)

#### Metric

p_final = Xval['Current_Mean_Price'] + p
yval_final = Xval['Current_Mean_Price'] + yval

np.sqrt(mean_squared_log_error(yval_final, p_final)) * 100

## Lasso

mdl = Lasso()
mdl.fit(Xtr, ytr)
p = mdl.predict(Xval)

#### Metric

p_final = Xval['Current_Mean_Price'] + p
yval_final = Xval['Current_Mean_Price'] + yval

np.sqrt(mean_squared_log_error(yval_final, p_final)) * 100

## Ridge

mdl = Ridge()
mdl.fit(Xtr, ytr)
p = mdl.predict(Xval)

#### Metric

p_final = Xval['Current_Mean_Price'] + p
yval_final = Xval['Current_Mean_Price'] + yval

np.sqrt(mean_squared_log_error(yval_final, p_final)) * 100

## BayesianRidge

mdl = BayesianRidge()
mdl.fit(Xtr, ytr)
p = mdl.predict(Xval)

#### Metric

p_final = Xval['Current_Mean_Price'] + p
yval_final = Xval['Current_Mean_Price'] + yval

np.sqrt(mean_squared_log_error(yval_final, p_final)) * 100
