==> 02-NoteBooks/04-FinancialEconomicData.ipynb <==
# Libraries

import pandas as pd
import os
import quandl

pd.set_option('display.max_columns', 200)

# Quandl Connection

quandl_api_key = os.environ.get('QUANDL_API_KEY')
quandl.ApiConfig.api_key = quandl_api_key

### Oil Price

wti_crude = quandl.get('EIA/PET_RWTC_D' , start_date='2004-05-03', end_date='2019-06-23')
wti_crude.rename(columns={'Value': 'Oil'}, inplace=True)

### Selic

selic = quandl.get('BCB/432' , start_date='2004-05-03', end_date='2019-06-23')
selic.rename(columns={'Value': 'Selic'}, inplace=True)

### Dollar

dollar = quandl.get('BCB/10813' , start_date='2004-05-03', end_date='2019-06-23')
dollar.rename(columns={'Value': 'Dollar'}, inplace=True)

### PIB BR

pib_BR = quandl.get('ODA/BRA_NGDPD' , start_date='2004-05-03', end_date='2019-06-23')
pib_BR.rename(columns={'Value': 'PIB_BR'}, inplace=True)

### IPCA

ipca = quandl.get('BCB/13522' , start_date='2004-05-03', end_date='2019-06-23')
ipca.rename(columns={'Value': 'IPCA'}, inplace=True)

### GOLD BR

gold = quandl.get('BCB/4' , start_date='2004-05-03', end_date='2019-06-23')
gold.rename(columns={'Value': 'Gold'}, inplace=True)

### CDI

cdi = quandl.get('BCB/4392' , start_date='2004-05-03', end_date='2019-06-23')
cdi.rename(columns={'Value': 'CDI'}, inplace=True)

# Join all the economic data into one DataFrame

economic_data = pd.concat([wti_crude, selic, dollar, pib_BR, ipca, gold, cdi], axis=1)

economic_data.head(10)

# Verification of Missing Values

economic_data.isnull().mean()

#### >>Economic indicators with a high incidence of null values will be discarded

# Statistical Summary

economic_data.describe()

# Drop columns with high incidence of null Values

economic_data.drop(['PIB_BR', 'IPCA', 'CDI'], axis=1, inplace=True)

economic_data.sample(10)

# Interpolation of values to fill the missing values

columns_with_null_values = economic_data.columns[economic_data.isnull().any()]

for col in columns_with_null_values:
    economic_data[col].interpolate(inplace=True)

economic_data.isnull().mean()

# Checking for missing days in the selected interval

economic_data['diff_Date'] = economic_data.index
economic_data['diff_Date'] = economic_data['diff_Date'].diff().shift(-1)

economic_data['diff_Date'].value_counts()

economic_data.drop('diff_Date', axis=1, inplace=True)

# Resample DataFrame to Week

col_names = economic_data.columns

economic_data_mean = economic_data.resample('W').mean()
economic_data_mean.columns = col_names + '_mean'

economic_data_median = economic_data.resample('W').median()
economic_data_median.columns = col_names + '_median'

economic_data_std = economic_data.resample('W').std()
economic_data_std.columns = col_names + '_std'

# Join all the resample metrics data into one DataFrame

economic_data = pd.concat([economic_data_mean, economic_data_median, economic_data_std], axis=1)

economic_data.head()

economic_data.to_csv('../01-Data/economic_data.csv', index=True)
