==> 02-NoteBooks/05-Features.ipynb <==
# Libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import warnings

warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', 200)

from sklearn.metrics import mean_squared_log_error

from lightgbm import LGBMRegressor
from sklearn.linear_model import LinearRegression

from sklearn.preprocessing import MinMaxScaler, StandardScaler, PowerTransformer

# Read Data

### Oil data

data = pd.read_csv('../01-Data/DataGas.csv', parse_dates=['Analysis_Date', 'Last_Day_of_Analyses_of_Week'])

data.head()

data.columns

### Economic Data

economic_data = pd.read_csv('../01-Data/economic_data.csv',  parse_dates=['Date'])

economic_data.head()

economic_data.columns

# Join the Economic Data with the oil data

data = data.merge(economic_data, left_on='Analysis_Date', right_on='Date')

# Train and Validation Split (Simple Holdout)

data_train = data[data['Last_Day_of_Analyses_of_Week'] < '2011-01-01']
data_valid = data[data['Last_Day_of_Analyses_of_Week'] >= '2011-01-01']

# New DataFrame for Train and Validation (Index: original Data)

df_train = pd.DataFrame(index=data_train.index)
df_valid  = pd.DataFrame(index=data_valid.index)

# Feature Engineering

MM_scaler = MinMaxScaler()
SS_scaler = StandardScaler()
pow_trans = PowerTransformer()


dataset_list = [(df_train, data_train), (df_valid, data_valid)]
var_one_hot = ['Macroregion', 'State']
regex = ['_mean', '_median', '_std']

def one_hot_encoding(model_dataset, sample_dataset, var):
    for label in sample_dataset[var].unique():
        model_dataset[var + '_' + label] = np.where(sample_dataset[var] == label, 1, 0)


for dataset in dataset_list:
    
    model_dataset, sample_dataset = dataset
    
    #Target: First Difference of Average Resale Price
    model_dataset['diff_Mean_Price'] = sample_dataset.groupby(['State'])['Mean_Price'].apply(lambda row: row.diff().shift(-1))
    
    ## Features
    # Current Mean Price
    model_dataset['Current_Mean_Price'] = sample_dataset['Mean_Price']
    
    # Mean Price Margin
    sample_dataset['Mean_Price_Margin'] = sample_dataset['Mean_Price_Margin'].astype(str)
    sample_dataset['Mean_Price_Margin'] = sample_dataset['Mean_Price_Margin'].replace('-', 0)
    sample_dataset['Mean_Price_Margin'] = sample_dataset['Mean_Price_Margin'].astype(float)
    model_dataset['Mean_Price_Margin'] = sample_dataset['Mean_Price_Margin'].astype(float)
    
    # Mean Price Margin with lag 1 week
    model_dataset['Mean_Price_Margin_lag_1_week'] = sample_dataset.groupby(['State'])['Mean_Price_Margin'].apply(lambda row: row.diff().shift(1))
    
    # Ratio Mean Price Margin / Mean Price Margin with lag 1 week
    model_dataset['Mean_Price_Margin_ratio'] = model_dataset['Mean_Price_Margin'] / model_dataset['Mean_Price_Margin_lag_1_week']
        
    #Seasonality
    model_dataset['month'] = sample_dataset['Last_Day_of_Analyses_of_Week'].dt.month
    model_dataset['day'] = sample_dataset['Last_Day_of_Analyses_of_Week'].dt.day
    model_dataset['dayofyear'] = sample_dataset['Last_Day_of_Analyses_of_Week'].dt.dayofyear
    model_dataset['year'] = sample_dataset['Last_Day_of_Analyses_of_Week'].dt.year
    #Encoding cyclic features
    columns = ['month', 'day', 'dayofyear', 'year']
    for col in columns:
        model_dataset[col + '_sin'] = np.sin((2*np.pi*model_dataset[col])/max(model_dataset[col]))
        model_dataset[col + '_cos'] = np.cos((2*np.pi*model_dataset[col])/max(model_dataset[col]))
         
    #Movel Average
    model_dataset['Movel_Average_Mean_Price_4_weeks'] = sample_dataset.groupby(['State'])['Mean_Price'].rolling(4).mean().reset_index(level=0, drop=True)
    
    #Economic Columns
    for reg in regex:
        economic_columns = economic_data.filter(regex=reg, axis=1).columns
        for col in economic_columns:
            model_dataset[col] = sample_dataset[col]
            # MinMaxScaler
            MM_scaler.fit(df_train[[col]])
            model_dataset[col + '_MM'] = MM_scaler.transform(model_dataset[[col]])
            # StandardScaler
            SS_scaler.fit(df_train[[col]])
            model_dataset[col + '_SS'] = SS_scaler.transform(model_dataset[[col]])
            # PowerTransformer
            pow_trans.fit(df_train[[col]])
            model_dataset[col + '_PW'] = pow_trans.transform(model_dataset[[col]])
            
    #Ratio Current_Mean_Price / Dollar
    model_dataset['Current_Mean_Price_Dollar_mean_ratio'] = model_dataset['Current_Mean_Price'] / model_dataset['Dollar_mean']
    #Ratio Current_Mean_Price / Oil
    model_dataset['Current_Mean_Price_Oil_mean_ratio'] = model_dataset['Current_Mean_Price'] / model_dataset['Oil_mean']
    
    
    #One Hot Encoding
    for var in var_one_hot:
        one_hot_encoding(model_dataset, sample_dataset, var)

# Drop Missing Values

df_train = df_train.dropna()
df_valid = df_valid.dropna()

# X, y Train and Validation Split

Xtr, ytr = df_train.drop(['diff_Mean_Price'], axis=1), df_train['diff_Mean_Price']
Xval, yval = df_valid.drop(['diff_Mean_Price'], axis=1), df_valid['diff_Mean_Price']

# Features Evaluation

### With One Hot Encoding

var_menor_erro = None
valor_menor_erro = 1000.

results = {'State_':{}, 'Macroregion_':{}}
one_hot_encode_params = ['State_', 'Macroregion_']

for param in one_hot_encode_params:
    param_exclude = list(filter(lambda x: x != param, one_hot_encode_params))[0]
    one_hot_exclude = list(Xtr.filter(regex=param_exclude).columns)
    encode_columns = list(Xtr.filter(regex=param).columns)
    to_drop = encode_columns + one_hot_exclude
    feature_columns = Xtr.drop(to_drop, axis=1).columns
    
    print(f'One Hot Encoding Param: {param}')
    
    for var in feature_columns:
        column_list = list(encode_columns)
        column_list.append(var)
        mdl = LGBMRegressor(num_leaves=2, min_data_in_leaf=250, n_jobs=-1, random_state=0, n_estimators=500)
        mdl.fit(Xtr[column_list], ytr)
        p = mdl.predict(Xval[column_list])

        p_final = Xval['Current_Mean_Price'] + p
        yval_final = Xval['Current_Mean_Price'] + yval

        erro = np.sqrt(mean_squared_log_error(yval_final, p_final)) * 100

        print("Variável: {} - Erro: {:.4f}\n".format(var, erro))

        if erro < valor_menor_erro:
            var_menor_erro = var
            valor_menor_erro = erro

        results[param][var] = erro
        
    print("Melhor Variável: {} - Erro: {:.4f} | One Hot Encoding {}\n\n".format(var_menor_erro, valor_menor_erro, param))
    
df_results = pd.DataFrame.from_dict(results, orient='index').T.rename(columns={param: param + 'result' for param in one_hot_encode_params})

### State Encode

baseline = 0.98587
df_results[df_results.iloc[:,0] < baseline].iloc[:,0].sort_values()

### Macroregion Encode

baseline = 0.98587
df_results[df_results.iloc[:,1] < baseline].iloc[:,1].sort_values()

### Without One Hot Encode

var_menor_erro = None
valor_menor_erro = 1000.

results = {}

for var in feature_columns:
    mdl = LGBMRegressor(num_leaves=2, min_data_in_leaf=250, n_jobs=-1, random_state=0, n_estimators=500)
    mdl.fit(Xtr[[var]], ytr)
    p = mdl.predict(Xval[[var]])

    p_final = Xval['Current_Mean_Price'] + p
    yval_final = Xval['Current_Mean_Price'] + yval

    erro = np.sqrt(mean_squared_log_error(yval_final, p_final)) * 100
    
    print("Variável: {} - Erro: {:.4f}\n".format(var, erro))
    
    if erro < valor_menor_erro:
        var_menor_erro = var
        valor_menor_erro = erro
        
    results[var] = erro
        
print("Melhor Variável: {} - Erro: {:.4f}\n".format(var_menor_erro, valor_menor_erro))

df_results = pd.DataFrame.from_dict(results, orient='index', columns=['result'])
baseline = 0.98587
df_results[df_results['result'] < baseline].sort_values('result')

### Model With Some Features and the State Encode

encode_columns = Xtr.filter(regex='State_').columns
enc_columns = list(encode_columns)
feat_columns = ['Gold_mean', 'Mean_Price_Margin_ratio', 'month_cos', 'Current_Mean_Price_Dollar_mean_ratio']
feat_columns = feat_columns + enc_columns

mdl = LGBMRegressor(num_leaves=2, min_data_in_leaf=250, n_jobs=-1, random_state=0, n_estimators=500)
mdl.fit(Xtr[feat_columns], ytr)
p = mdl.predict(Xval[feat_columns])

p_final = Xval['Current_Mean_Price'] + p
yval_final = Xval['Current_Mean_Price'] + yval

np.sqrt(mean_squared_log_error(yval_final, p_final)) * 100

### Model With Some Features and the Macroregion Encode

encode_columns = Xtr.filter(regex='Macroregion_').columns
enc_columns = list(encode_columns)
feat_columns = ['Gold_mean', 'Mean_Price_Margin_ratio', 'month_cos', 'Current_Mean_Price_Dollar_mean_ratio']
feat_columns = feat_columns + enc_columns

mdl = LGBMRegressor(num_leaves=2, min_data_in_leaf=250, n_jobs=-1, random_state=0, n_estimators=500)
mdl.fit(Xtr[feat_columns], ytr)
p = mdl.predict(Xval[feat_columns])

p_final = Xval['Current_Mean_Price'] + p
yval_final = Xval['Current_Mean_Price'] + yval

np.sqrt(mean_squared_log_error(yval_final, p_final)) * 100

### Model With Some Features and without Encode

feat_columns = ['Gold_mean', 'Mean_Price_Margin_ratio', 'month_cos', 'Current_Mean_Price_Dollar_mean_ratio']


mdl = LGBMRegressor(num_leaves=2, min_data_in_leaf=250, n_jobs=-1, random_state=0, n_estimators=500)
mdl.fit(Xtr[feat_columns], ytr)
p = mdl.predict(Xval[feat_columns])

p_final = Xval['Current_Mean_Price'] + p
yval_final = Xval['Current_Mean_Price'] + yval

np.sqrt(mean_squared_log_error(yval_final, p_final)) * 100
